Quick fix from double to float conversion. Need to switch plume to float.

+++ src/plume/GLE_Solver.cpp
@@ -57,8 +57,15 @@ void GLE_Solver_CPU::solve(Particle *p, double &par_dt, TURBGeneralData *TGD, PL
                             p->nuT,
                             p->CoEps);
 
-  mat3sym tau = { txx, txy, txz, tyy, tyz, tzz };
-  vec3 flux_div = { flux_div_x, flux_div_y, flux_div_z };
+  mat3sym tau = { static_cast<float>(txx),
+                  static_cast<float>(txy),
+                  static_cast<float>(txz),
+                  static_cast<float>(tyy),
+                  static_cast<float>(tyz),
+                  static_cast<float>(tzz) };
+  vec3 flux_div = { static_cast<float>(flux_div_x),
+                    static_cast<float>(flux_div_y),
+                    static_cast<float>(flux_div_z) };
 
   // mat3sym tau = { 0.0, 0.0, 0.0, 0.0, 0.0, 0.0 };
   // vec3 flux_div = { 0.0, 0.0, 0.0 };
@@ -82,33 +89,33 @@ void GLE_Solver_CPU::solve(Particle *p, double &par_dt, TURBGeneralData *TGD, PL
   // these are the random numbers for each direction
 
 #ifdef _OPENMP
-  vec3 vRandn = { PGD->threadRNG[omp_get_thread_num()]->norRan(),
-                  PGD->threadRNG[omp_get_thread_num()]->norRan(),
-                  PGD->threadRNG[omp_get_thread_num()]->norRan() };
+  vec3 vRandn = { static_cast<float>(PGD->threadRNG[omp_get_thread_num()]->norRan()),
+                  static_cast<float>(PGD->threadRNG[omp_get_thread_num()]->norRan()),
+                  static_cast<float>(PGD->threadRNG[omp_get_thread_num()]->norRan()) };
 #else
   vec3 vRandn = { PGD->RNG->norRan(), PGD->RNG->norRan(), PGD->RNG->norRan() };
 #endif
 
   // now calculate a bunch of values for the current particle
   // calculate the time derivative of the stress tensor: (tau_current - tau_old)/dt
-  mat3sym tau_ddt = { (tau._11 - p->tau._11) / par_dt,
-                      (tau._12 - p->tau._12) / par_dt,
-                      (tau._13 - p->tau._13) / par_dt,
-                      (tau._22 - p->tau._22) / par_dt,
-                      (tau._23 - p->tau._23) / par_dt,
-                      (tau._33 - p->tau._33) / par_dt };
+  mat3sym tau_ddt = { static_cast<float>((tau._11 - p->tau._11) / par_dt),
+                      static_cast<float>((tau._12 - p->tau._12) / par_dt),
+                      static_cast<float>((tau._13 - p->tau._13) / par_dt),
+                      static_cast<float>((tau._22 - p->tau._22) / par_dt),
+                      static_cast<float>((tau._23 - p->tau._23) / par_dt),
+                      static_cast<float>((tau._33 - p->tau._33) / par_dt) };
 
   // now calculate and set the A and b matrices for an Ax = b
   // A = -I + 0.5*(-CoEps*L + dTdt*L )*par_dt;
-  mat3 A = { -1.0 + 0.50 * (-p->CoEps * L._11 + L._11 * tau_ddt._11 + L._12 * tau_ddt._12 + L._13 * tau_ddt._13) * par_dt,
-             -0.0 + 0.50 * (-p->CoEps * L._12 + L._12 * tau_ddt._11 + L._22 * tau_ddt._12 + L._23 * tau_ddt._13) * par_dt,
-             -0.0 + 0.50 * (-p->CoEps * L._13 + L._13 * tau_ddt._11 + L._23 * tau_ddt._12 + L._33 * tau_ddt._13) * par_dt,
-             -0.0 + 0.50 * (-p->CoEps * L._12 + L._11 * tau_ddt._12 + L._12 * tau_ddt._22 + L._13 * tau_ddt._23) * par_dt,
-             -1.0 + 0.50 * (-p->CoEps * L._22 + L._12 * tau_ddt._12 + L._22 * tau_ddt._22 + L._23 * tau_ddt._23) * par_dt,
-             -0.0 + 0.50 * (-p->CoEps * L._23 + L._13 * tau_ddt._12 + L._23 * tau_ddt._22 + L._33 * tau_ddt._23) * par_dt,
-             -0.0 + 0.50 * (-p->CoEps * L._13 + L._11 * tau_ddt._13 + L._12 * tau_ddt._23 + L._13 * tau_ddt._33) * par_dt,
-             -0.0 + 0.50 * (-p->CoEps * L._23 + L._12 * tau_ddt._13 + L._22 * tau_ddt._23 + L._23 * tau_ddt._33) * par_dt,
-             -1.0 + 0.50 * (-p->CoEps * L._33 + L._13 * tau_ddt._13 + L._23 * tau_ddt._23 + L._33 * tau_ddt._33) * par_dt };
+  mat3 A = { static_cast<float>(-1.0 + 0.50 * (-p->CoEps * L._11 + L._11 * tau_ddt._11 + L._12 * tau_ddt._12 + L._13 * tau_ddt._13) * par_dt),
+             static_cast<float>(-0.0 + 0.50 * (-p->CoEps * L._12 + L._12 * tau_ddt._11 + L._22 * tau_ddt._12 + L._23 * tau_ddt._13) * par_dt),
+             static_cast<float>(-0.0 + 0.50 * (-p->CoEps * L._13 + L._13 * tau_ddt._11 + L._23 * tau_ddt._12 + L._33 * tau_ddt._13) * par_dt),
+             static_cast<float>(-0.0 + 0.50 * (-p->CoEps * L._12 + L._11 * tau_ddt._12 + L._12 * tau_ddt._22 + L._13 * tau_ddt._23) * par_dt),
+             static_cast<float>(-1.0 + 0.50 * (-p->CoEps * L._22 + L._12 * tau_ddt._12 + L._22 * tau_ddt._22 + L._23 * tau_ddt._23) * par_dt),
+             static_cast<float>(-0.0 + 0.50 * (-p->CoEps * L._23 + L._13 * tau_ddt._12 + L._23 * tau_ddt._22 + L._33 * tau_ddt._23) * par_dt),
+             static_cast<float>(-0.0 + 0.50 * (-p->CoEps * L._13 + L._11 * tau_ddt._13 + L._12 * tau_ddt._23 + L._13 * tau_ddt._33) * par_dt),
+             static_cast<float>(-0.0 + 0.50 * (-p->CoEps * L._23 + L._12 * tau_ddt._13 + L._22 * tau_ddt._23 + L._23 * tau_ddt._33) * par_dt),
+             static_cast<float>(-1.0 + 0.50 * (-p->CoEps * L._33 + L._13 * tau_ddt._13 + L._23 * tau_ddt._23 + L._33 * tau_ddt._33) * par_dt) };
 
   /*double A_11 = -1.0 + 0.50 * (-p->CoEps * lxx + lxx * dtxxdt + lxy * dtxydt + lxz * dtxzdt) * par_dt;
   double A_12 = 0.50 * (-p->CoEps * lxy + lxy * dtxxdt + lyy * dtxydt + lyz * dtxzdt) * par_dt;
@@ -123,9 +130,9 @@ void GLE_Solver_CPU::solve(Particle *p, double &par_dt, TURBGeneralData *TGD, PL
   double A_33 = -1.0 + 0.50 * (-p->CoEps * lzz + lxz * dtxzdt + lyz * dtyzdt + lzz * dtzzdt) * par_dt;*/
 
   // b = -vectFluct - 0.5*vecFluxDiv*par_dt - sqrt(CoEps*par_dt)*vecRandn;
-  vec3 b = { -p->velFluct_old._1 - 0.50 * flux_div._1 * par_dt - std::sqrt(p->CoEps * par_dt) * vRandn._1,
-             -p->velFluct_old._2 - 0.50 * flux_div._2 * par_dt - std::sqrt(p->CoEps * par_dt) * vRandn._2,
-             -p->velFluct_old._3 - 0.50 * flux_div._3 * par_dt - std::sqrt(p->CoEps * par_dt) * vRandn._3 };
+  vec3 b = { static_cast<float>(-p->velFluct_old._1 - 0.50 * flux_div._1 * par_dt - std::sqrt(p->CoEps * par_dt) * vRandn._1),
+             static_cast<float>(-p->velFluct_old._2 - 0.50 * flux_div._2 * par_dt - std::sqrt(p->CoEps * par_dt) * vRandn._2),
+             static_cast<float>(-p->velFluct_old._3 - 0.50 * flux_div._3 * par_dt - std::sqrt(p->CoEps * par_dt) * vRandn._3) };
 
   /*double b_11 = -p->uFluct_old - 0.50 * flux_div_x * par_dt - std::sqrt(p->CoEps * par_dt) * xRandn;
   double b_21 = -p->vFluct_old - 0.50 * flux_div_y * par_dt - std::sqrt(p->CoEps * par_dt) * yRandn;
